Add pipeline information to Pipeline page
