+++ src/components/configuration/granularityMappingConfig.js
