+++ src/components/configuration/granuralityConfig.js
