+++ src/components/configuration/schemaConfig.js
@@ -1,69 +1,67 @@
 import React from 'react';
-import axios from 'axios';
 import Api from '../api';
 
-class BulkIngest extends React.Component {
+class SchemaConfig extends React.Component {
     state = {
         featureName: "",
-        sourceType:"",
-        sourceFormat:"",
-        transformation: [{ attributeName: "", transformation: ""}],
-        geomConfigurations: [{ geomFormat: "POINT", dataSource: "", featureId: "" }],
-        dataSources:[{source:""}],
+        attributes: [{ attribute: "", attributeName: "", attributeType: "", derived: "" }],
+        configurations: [{ geom: "POINT", dtg: "YYYYMMDD", featureId: "" }],
+        features: [{ featureName: "", attributes: [], configurations: [] }],
         postingFeatures: [],
-        errorMsg: { featureName: "", sourceType:"", sourceFormat:"", transformation: "", geomConfiguration:"" },
+        errorMsg: { featureName: "", atttributes: "", configurations: "" },
         response: ""
     }
 
     api = new Api();
 
+
     handleChange = (e) => {
         let errorMsg = this.state.errorMsg
         let id = e.target.dataset.id
-        if (["attributeName", "attributeType"].includes(e.target.id)) {
-            let transformation = [...this.state.transformation]
+        if (["attribute", "attributeName", "attributeType", "derived"].includes(e.target.id)) {
+            let attributes = [...this.state.attributes]
             console.log("data set id:", e.target.dataset.id)
-            transformation[e.target.dataset.id][e.target.id] = e.target.value.toUpperCase()
-            this.setState({ transformation }, () => {
+            attributes[e.target.dataset.id][e.target.id] = e.target.value.toUpperCase()
+            this.setState({ attributes }, () => {
                 let err = '';
-                if (!this.state.transformation[id]["attributeName"] ||
-                    !this.state.transformation[id]["transformation"] ) {
+                if (!this.state.attributes[id]["attribute"] || !this.state.attributes[id]["attributeName"] ||
+                    !this.state.attributes[id]["attributeType"] || !this.state.attributes[id]["derived"]) {
                     err = "Attribute fieldS can not be empty";
-                    errorMsg["transformation"] = err
+                    errorMsg["atttributes"] = err
                     this.setState({ errorMsg });
                 } else {
                     err = "";
-                    errorMsg["transformation"] = err
+                    errorMsg["atttributes"] = err
                     this.setState({ errorMsg });
                 }
                 //console.log(this.state.attributes)
             })
-        }else if (["geomFormat", "dataSource", "featureId"].includes(e.target.id)) {
-                     console.log("configuration change")
-                     let geomConfigurations = [...this.state.configurations]
-                     let error = ""
-                     errorMsg["geomConfigurations"] = error
-                     this.setState({ errorMsg })
-                     geomConfigurations[0][e.target.id] = e.target.value.toUpperCase()
-                     this.setState({ geomConfigurations }, () => {
-                         let err = '';
-                         if (!this.state.geomConfigurations[0]["featureId"]) {
-                             err = "FeatureId field can not be empty";
-                             errorMsg["geomConfigurations"] = err
-                             this.setState({ errorMsg });
-                         } else {
-                             err = "";
-                             errorMsg["geomConfigurations"] = err
-                             this.setState({ errorMsg });
-                         }
-                     })
-                 }
+        } else if (["geom", "dtg", "featureId"].includes(e.target.id)) {
+            console.log("configuration change")
+            let configurations = [...this.state.configurations]
+            let error = ""
+            errorMsg["configurations"] = error
+            this.setState({ errorMsg })
+            configurations[0][e.target.id] = e.target.value.toUpperCase()
+            this.setState({ configurations }, () => {
+                let err = '';
+                if (!this.state.configurations[0]["featureId"]) {
+                    err = "FeatureId field can not be empty";
+                    errorMsg["configurations"] = err
+                    this.setState({ errorMsg });
+                } else {
+                    err = "";
+                    errorMsg["configurations"] = err
+                    this.setState({ errorMsg });
+                }
+            })
+        }
         else {
             let error = ""
             if (!e.target.value) {
                 error = `${e.target.name} field cannot be empty`
             }
-            errorMsg[e.target.id] = error
+            errorMsg["featureName"] = error
             this.setState({ errorMsg });
             this.setState({ [e.target.name]: e.target.value.toUpperCase() })
         }
@@ -73,35 +71,21 @@ class BulkIngest extends React.Component {
 
     addAttribute = (e) => {
         this.setState((prevState) => ({
-            transformation: [...prevState.transformation, { attributeName: "", transformation: ""}]
+            attributes: [...prevState.attributes, { attribute: "", attributeName: "", attributeType: "", derived: "" }]
         }));
     }
     removeAttribute = (e) => {
         let errorMsg = this.state.errorMsg
-        var arrayTransformation = this.state.transformation;
-        if (arrayTransformation.length > 0) {
-            arrayTransformation.splice(-1, 1)
+        var arrayAttribute = this.state.attributes;
+        var arrayRecord = this.state.records;
+        if (arrayAttribute.length > 0) {
+            arrayAttribute.splice(-1, 1)
+            arrayRecord.splice(-1, 1)
         }
-        errorMsg["transformation"] = ""
+        errorMsg["atttributes"] = ""
         this.setState((prevState) => ({
-            transformation: arrayTransformation,
-            errorMsg: errorMsg
-        }));
-    }
-    addSource = (e) => {
-        this.setState((prevState) => ({
-            dataSources: [...prevState.dataSources, {source:""}]
-        }));
-    }
-    removeSource = (e) => {
-        let errorMsg = this.state.errorMsg
-        var arraySources = this.state.dataSources;
-        if (arraySources.length > 0) {
-            arraySources.splice(-1, 1)
-        }
-        errorMsg["dataSources"] = ""
-        this.setState((prevState) => ({
-            dataSources: arraySources,
+            attributes: arrayAttribute,
+            records: arrayRecord,
             errorMsg: errorMsg
         }));
     }
@@ -121,9 +105,30 @@ class BulkIngest extends React.Component {
             this.setState({ errorMsg });
         } else if (this.state.errorMsg[1]) {
         } else {
+            // console.log(this.state.errorMsg)
+            // var id = this.state.features.length
+
+            // let features = [...this.state.features]
+            // features[id - 1]["featureName"] = this.state.featureName
+            // features[id - 1]["attributes"] = this.state.attributes
+            // features[id - 1]["records"] = this.state.records
+            // features[id - 1]["configurations"] = this.state.configurations
+            // this.setState({ features }, () => {
+            //     console.log(this.state.features)
+
+            // })
+            // this.setState((prevState) => ({
+            //     features: [...prevState.features, { featureName: "", attributes: [], records: [], configurations: [] }],
+            //     attributes: [{ attribute: "", attributeName: "", attributeType: "", derived: "" }],
+            //     records: [{ attributeName: "", columnNumber: "" }],
+            //     configurations: [{ geom: "POINT", dtg: "YYYYMMDD", featureId: "" }],
+            //     featureName: ""
+            // }))
+
             let addedFeature = {
                 featureName: this.state.featureName,
-                attributes: this.state.transformation,
+                attributes: this.state.attributes,
+                records: this.state.records,
                 configurations: this.state.configurations
             }
 
@@ -134,8 +139,9 @@ class BulkIngest extends React.Component {
             });
 
             this.setState((prevState) => ({
-                features: [{ featureName: "", attributes: [], configurations: [] }],
+                features: [{ featureName: "", attributes: [], records: [], configurations: [] }],
                 attributes: [{ attribute: "", attributeName: "", attributeType: "", derived: "" }],
+                records: [{ attributeName: "", columnNumber: "" }],
                 configurations: [{ geom: "POINT", dtg: "YYYYMMDD", featureId: "" }],
                 featureName: ""
             }))
@@ -145,7 +151,7 @@ class BulkIngest extends React.Component {
             Array.from(document.querySelectorAll("input")).forEach(
                 input => (input.value = "")
             );
-            errorMsg = { featureName: "", atttributes: "", configurations: "" }
+            errorMsg = { featureName: "", atttributes: "", records: "", configurations: "" }
             this.setState({ errorMsg })
         }
     }
@@ -156,7 +162,7 @@ class BulkIngest extends React.Component {
         }
         this.setState((prevState) => ({
             features: arrayFeature,
-            errorMsg: { featureName: "", atttributes: "", configurations: "" }
+            errorMsg: { featureName: "", atttributes: "", records: "", configurations: "" }
         }), () => {
             console.log(this.state.features)
         });
@@ -182,33 +188,38 @@ class BulkIngest extends React.Component {
         e.preventDefault()
     }
     render() {
-        let { transformation, dataSources} = this.state
+        let { attributes, features } = this.state
         return (
             <div className="w3-border">
                 <form className="w3-container" onSubmit={this.handleSubmit} onChange={this.handleChange}>
                     <h6>
                         <label>Feature Name</label>
                         <input className="w3-input" type="text" name="featureName"></input>
-                        <label>Source Type</label>
-                        <input className="w3-input" type="text" name="sourceType"></input>
-                        <label>Source Format</label>
-                        <input className="w3-input" type="text" name="sourceFormat"></input>
                         <div className="h7">{this.state.errorMsg["featureName"]}</div>
                         <br />
-                        <label>Transformation</label>
+                        <label>Attributes</label>
                         <br /><br />
                         {
-                            transformation.map((val, idx) => {
-                                let nameId = `name-${idx}`, typeId = `type-${idx}`
+                            attributes.map((val, idx) => {
+                                let attributeId = `attribute-${idx}`, nameId = `name-${idx}`, typeId = `type-${idx}`, derivedId = `derive-${idx}`
                                 return (
                                     <div key={idx} className="row w3-panel w3-border">
-                                        <label htmlFor={nameId} className="col-25">Attribute Name</label>
+                                        <label htmlFor={attributeId} className="col-25">Attribute #{idx + 1}</label>
+                                        <input
+                                            type="text"
+                                            name={attributeId}
+                                            data-id={idx}
+                                            id="attribute"
+                                            value={attributes[idx].atr}
+                                            className="col-75"
+                                        />
+                                        <label htmlFor={nameId} className="col-25">Name</label>
                                         <input
                                             type="text"
                                             name={nameId}
                                             data-id={idx}
                                             id="attributeName"
-                                            value={transformation[idx].ame}
+                                            value={attributes[idx].ame}
                                             className="col-75"
                                         />
                                         <label htmlFor={typeId} className="col-25">Type</label>
@@ -217,7 +228,16 @@ class BulkIngest extends React.Component {
                                             name={typeId}
                                             data-id={idx}
                                             id="attributeType"
-                                            value={transformation[idx].type}
+                                            value={attributes[idx].type}
+                                            className="col-75 "
+                                        />
+                                        <label htmlFor={typeId} className="col-25">Derived</label>
+                                        <input
+                                            type="text"
+                                            name={derivedId}
+                                            data-id={idx}
+                                            id="derived"
+                                            value={attributes[idx].derv}
                                             className="col-75 "
                                         />
                                         <br />
@@ -225,28 +245,15 @@ class BulkIngest extends React.Component {
                                 )
                             })
                         }
-                        <div className="h7">{this.state.errorMsg["transformation"]}</div>
+                        <div className="h7">{this.state.errorMsg["atttributes"]}</div>
                         <br />
                         <button className="w3-button w3-circle w3-teal" onClick={this.removeAttribute}>-</button>
                         <button className="w3-button w3-circle w3-teal" onClick={this.addAttribute}>+</button>
                         <br /><br />
-                        <label>Data Sources</label>
-                        {
-                            dataSources.map((val, idx) => {
-                                let inputId =`${idx}`
-                                return(
-                                    <input className="w3-input" dataId={inputId} type="text" name="dataSources"></input>
-                                )
-                        })
-                        }
+                        <label>Configuration</label>
                         <br />
-                        <button className="w3-button w3-circle w3-teal" onClick={this.removeSource}>-</button>
-                        <button className="w3-button w3-circle w3-teal" onClick={this.addSource}>+</button>
-                        <br /><br />
-                        <label>Geom Configuration</label>
-                        <br />
-                        <div className="row" id="geomConfiguration">
-                            <label className="col-50">Geom Format</label>
+                        <div className="row" id="configuration">
+                            <label className="col-50">GeomFormat</label>
                             <div className="col-50">
                                 <select id="geom" name="geom">
                                     <option value="point">Point</option>
@@ -254,14 +261,18 @@ class BulkIngest extends React.Component {
                                     <option value="multipolygon">Multi Polygon</option>
                                 </select>
                             </div>
-                            <label className="col-50">Data Source</label>
-                            <input className="col-25" type="text" id="dataSource"></input>
-                            <div className="h7">{this.state.errorMsg["geomConfigurations"]}</div>
-                            <label className="col-50">Feature ID</label>
+                            <label className="col-50">DateTimeFormat</label>
+                            <div className="col-50">
+                                <select id="dtg" name="dtg" >
+                                    <option value="yyyyMMdd">yyyy/mm/dd</option>
+                                    <option value="ddMMyyyy">dd/mm/yyyy</option>
+                                    <option value="MMddyyyy">mmm/dd/yyyy</option>
+                                </select>
+                            </div>
+                            <label className="col-50">FeatureID</label>
                             <input className="col-25" type="text" id="featureId"></input>
-                            <div className="h7">{this.state.errorMsg["geomConfigurations"]}</div>
+                            <div className="h7">{this.state.errorMsg["configurations"]}</div>
                         </div>
-
                     </h6>
                     <button className="w3-btn w3-white w3-border w3-border-red w3-round-large" onClick={this.removeFeature}>Remove</button>
                     <button className="w3-btn w3-white w3-border w3-border-green w3-round-large" onClick={this.addFeature}>Add Feature</button>
@@ -274,4 +285,4 @@ class BulkIngest extends React.Component {
     }
 }
 
-export default BulkIngest;
+export default SchemaConfig;
