+++ src/components/configuration/single/bulkIngest.js
@@ -1,58 +1,59 @@
 import React from 'react';
 import axios from 'axios';
-import Api from './api';
+import Api from '../../api';
 
-class SchemaConfig extends React.Component {
+class BulkIngest extends React.Component {
     state = {
         featureName: "",
-        attributes: [{ attribute: "", attributeName: "", attributeType: "", derived: "" }],
-        configurations: [{ geom: "POINT", dtg: "YYYYMMDD", featureId: "" }],
-        features: [{ featureName: "", attributes: [], configurations: [] }],
+        sourceType: "",
+        sourceFormat: "",
+        transformation: [{ attributeName: "", transformation: "" }],
+        geomConfigurations: [{ geomFormat: "POINT", dataSource: "", featureId: "" }],
+        dataSources: [{ source: "" }],
         postingFeatures: [],
-        errorMsg: { featureName: "", atttributes: "", configurations: "" },
+        errorMsg: { featureName: "", sourceType: "", sourceFormat: "", transformation: "", geomConfiguration: "" },
         response: ""
     }
 
     api = new Api();
 
-
     handleChange = (e) => {
         let errorMsg = this.state.errorMsg
         let id = e.target.dataset.id
-        if (["attribute", "attributeName", "attributeType", "derived"].includes(e.target.id)) {
-            let attributes = [...this.state.attributes]
+        if (["attributeName", "attributeType"].includes(e.target.id)) {
+            let transformation = [...this.state.transformation]
             console.log("data set id:", e.target.dataset.id)
-            attributes[e.target.dataset.id][e.target.id] = e.target.value.toUpperCase()
-            this.setState({ attributes }, () => {
+            transformation[e.target.dataset.id][e.target.id] = e.target.value.toUpperCase()
+            this.setState({ transformation }, () => {
                 let err = '';
-                if (!this.state.attributes[id]["attribute"] || !this.state.attributes[id]["attributeName"] ||
-                    !this.state.attributes[id]["attributeType"] || !this.state.attributes[id]["derived"]) {
+                if (!this.state.transformation[id]["attributeName"] ||
+                    !this.state.transformation[id]["transformation"]) {
                     err = "Attribute fieldS can not be empty";
-                    errorMsg["atttributes"] = err
+                    errorMsg["transformation"] = err
                     this.setState({ errorMsg });
                 } else {
                     err = "";
-                    errorMsg["atttributes"] = err
+                    errorMsg["transformation"] = err
                     this.setState({ errorMsg });
                 }
                 //console.log(this.state.attributes)
             })
-        } else if (["geom", "dtg", "featureId"].includes(e.target.id)) {
+        } else if (["geomFormat", "dataSource", "featureId"].includes(e.target.id)) {
             console.log("configuration change")
-            let configurations = [...this.state.configurations]
+            let geomConfigurations = [...this.state.configurations]
             let error = ""
-            errorMsg["configurations"] = error
+            errorMsg["geomConfigurations"] = error
             this.setState({ errorMsg })
-            configurations[0][e.target.id] = e.target.value.toUpperCase()
-            this.setState({ configurations }, () => {
+            geomConfigurations[0][e.target.id] = e.target.value.toUpperCase()
+            this.setState({ geomConfigurations }, () => {
                 let err = '';
-                if (!this.state.configurations[0]["featureId"]) {
+                if (!this.state.geomConfigurations[0]["featureId"]) {
                     err = "FeatureId field can not be empty";
-                    errorMsg["configurations"] = err
+                    errorMsg["geomConfigurations"] = err
                     this.setState({ errorMsg });
                 } else {
                     err = "";
-                    errorMsg["configurations"] = err
+                    errorMsg["geomConfigurations"] = err
                     this.setState({ errorMsg });
                 }
             })
@@ -62,7 +63,7 @@ class SchemaConfig extends React.Component {
             if (!e.target.value) {
                 error = `${e.target.name} field cannot be empty`
             }
-            errorMsg["featureName"] = error
+            errorMsg[e.target.id] = error
             this.setState({ errorMsg });
             this.setState({ [e.target.name]: e.target.value.toUpperCase() })
         }
@@ -72,21 +73,35 @@ class SchemaConfig extends React.Component {
 
     addAttribute = (e) => {
         this.setState((prevState) => ({
-            attributes: [...prevState.attributes, { attribute: "", attributeName: "", attributeType: "", derived: "" }]
+            transformation: [...prevState.transformation, { attributeName: "", transformation: "" }]
         }));
     }
     removeAttribute = (e) => {
         let errorMsg = this.state.errorMsg
-        var arrayAttribute = this.state.attributes;
-        var arrayRecord = this.state.records;
-        if (arrayAttribute.length > 0) {
-            arrayAttribute.splice(-1, 1)
-            arrayRecord.splice(-1, 1)
+        var arrayTransformation = this.state.transformation;
+        if (arrayTransformation.length > 0) {
+            arrayTransformation.splice(-1, 1)
         }
-        errorMsg["atttributes"] = ""
+        errorMsg["transformation"] = ""
         this.setState((prevState) => ({
-            attributes: arrayAttribute,
-            records: arrayRecord,
+            transformation: arrayTransformation,
+            errorMsg: errorMsg
+        }));
+    }
+    addSource = (e) => {
+        this.setState((prevState) => ({
+            dataSources: [...prevState.dataSources, { source: "" }]
+        }));
+    }
+    removeSource = (e) => {
+        let errorMsg = this.state.errorMsg
+        var arraySources = this.state.dataSources;
+        if (arraySources.length > 0) {
+            arraySources.splice(-1, 1)
+        }
+        errorMsg["dataSources"] = ""
+        this.setState((prevState) => ({
+            dataSources: arraySources,
             errorMsg: errorMsg
         }));
     }
@@ -106,30 +121,9 @@ class SchemaConfig extends React.Component {
             this.setState({ errorMsg });
         } else if (this.state.errorMsg[1]) {
         } else {
-            // console.log(this.state.errorMsg)
-            // var id = this.state.features.length
-
-            // let features = [...this.state.features]
-            // features[id - 1]["featureName"] = this.state.featureName
-            // features[id - 1]["attributes"] = this.state.attributes
-            // features[id - 1]["records"] = this.state.records
-            // features[id - 1]["configurations"] = this.state.configurations
-            // this.setState({ features }, () => {
-            //     console.log(this.state.features)
-
-            // })
-            // this.setState((prevState) => ({
-            //     features: [...prevState.features, { featureName: "", attributes: [], records: [], configurations: [] }],
-            //     attributes: [{ attribute: "", attributeName: "", attributeType: "", derived: "" }],
-            //     records: [{ attributeName: "", columnNumber: "" }],
-            //     configurations: [{ geom: "POINT", dtg: "YYYYMMDD", featureId: "" }],
-            //     featureName: ""
-            // }))
-
             let addedFeature = {
                 featureName: this.state.featureName,
-                attributes: this.state.attributes,
-                records: this.state.records,
+                attributes: this.state.transformation,
                 configurations: this.state.configurations
             }
 
@@ -140,9 +134,8 @@ class SchemaConfig extends React.Component {
             });
 
             this.setState((prevState) => ({
-                features: [{ featureName: "", attributes: [], records: [], configurations: [] }],
+                features: [{ featureName: "", attributes: [], configurations: [] }],
                 attributes: [{ attribute: "", attributeName: "", attributeType: "", derived: "" }],
-                records: [{ attributeName: "", columnNumber: "" }],
                 configurations: [{ geom: "POINT", dtg: "YYYYMMDD", featureId: "" }],
                 featureName: ""
             }))
@@ -152,7 +145,7 @@ class SchemaConfig extends React.Component {
             Array.from(document.querySelectorAll("input")).forEach(
                 input => (input.value = "")
             );
-            errorMsg = { featureName: "", atttributes: "", records: "", configurations: "" }
+            errorMsg = { featureName: "", atttributes: "", configurations: "" }
             this.setState({ errorMsg })
         }
     }
@@ -163,7 +156,7 @@ class SchemaConfig extends React.Component {
         }
         this.setState((prevState) => ({
             features: arrayFeature,
-            errorMsg: { featureName: "", atttributes: "", records: "", configurations: "" }
+            errorMsg: { featureName: "", atttributes: "", configurations: "" }
         }), () => {
             console.log(this.state.features)
         });
@@ -189,38 +182,33 @@ class SchemaConfig extends React.Component {
         e.preventDefault()
     }
     render() {
-        let { attributes, features } = this.state
+        let { transformation, dataSources } = this.state
         return (
             <div className="w3-border">
                 <form className="w3-container" onSubmit={this.handleSubmit} onChange={this.handleChange}>
                     <h6>
                         <label>Feature Name</label>
                         <input className="w3-input" type="text" name="featureName"></input>
+                        <label>Source Type</label>
+                        <input className="w3-input" type="text" name="sourceType"></input>
+                        <label>Source Format</label>
+                        <input className="w3-input" type="text" name="sourceFormat"></input>
                         <div className="h7">{this.state.errorMsg["featureName"]}</div>
                         <br />
-                        <label>Attributes</label>
+                        <label>Transformation</label>
                         <br /><br />
                         {
-                            attributes.map((val, idx) => {
-                                let attributeId = `attribute-${idx}`, nameId = `name-${idx}`, typeId = `type-${idx}`, derivedId = `derive-${idx}`
+                            transformation.map((val, idx) => {
+                                let nameId = `name-${idx}`, typeId = `type-${idx}`
                                 return (
                                     <div key={idx} className="row w3-panel w3-border">
-                                        <label htmlFor={attributeId} className="col-25">Attribute #{idx + 1}</label>
-                                        <input
-                                            type="text"
-                                            name={attributeId}
-                                            data-id={idx}
-                                            id="attribute"
-                                            value={attributes[idx].atr}
-                                            className="col-75"
-                                        />
-                                        <label htmlFor={nameId} className="col-25">Name</label>
+                                        <label htmlFor={nameId} className="col-25">Attribute Name</label>
                                         <input
                                             type="text"
                                             name={nameId}
                                             data-id={idx}
                                             id="attributeName"
-                                            value={attributes[idx].ame}
+                                            value={transformation[idx].ame}
                                             className="col-75"
                                         />
                                         <label htmlFor={typeId} className="col-25">Type</label>
@@ -229,16 +217,7 @@ class SchemaConfig extends React.Component {
                                             name={typeId}
                                             data-id={idx}
                                             id="attributeType"
-                                            value={attributes[idx].type}
-                                            className="col-75 "
-                                        />
-                                        <label htmlFor={typeId} className="col-25">Derived</label>
-                                        <input
-                                            type="text"
-                                            name={derivedId}
-                                            data-id={idx}
-                                            id="derived"
-                                            value={attributes[idx].derv}
+                                            value={transformation[idx].type}
                                             className="col-75 "
                                         />
                                         <br />
@@ -246,15 +225,28 @@ class SchemaConfig extends React.Component {
                                 )
                             })
                         }
-                        <div className="h7">{this.state.errorMsg["atttributes"]}</div>
+                        <div className="h7">{this.state.errorMsg["transformation"]}</div>
                         <br />
                         <button className="w3-button w3-circle w3-teal" onClick={this.removeAttribute}>-</button>
                         <button className="w3-button w3-circle w3-teal" onClick={this.addAttribute}>+</button>
                         <br /><br />
-                        <label>Configuration</label>
+                        <label>Data Sources</label>
+                        {
+                            dataSources.map((val, idx) => {
+                                let inputId = `${idx}`
+                                return (
+                                    <input className="w3-input" dataId={inputId} type="text" name="dataSources"></input>
+                                )
+                            })
+                        }
                         <br />
-                        <div className="row" id="configuration">
-                            <label className="col-50">GeomFormat</label>
+                        <button className="w3-button w3-circle w3-teal" onClick={this.removeSource}>-</button>
+                        <button className="w3-button w3-circle w3-teal" onClick={this.addSource}>+</button>
+                        <br /><br />
+                        <label>Geom Configuration</label>
+                        <br />
+                        <div className="row" id="geomConfiguration">
+                            <label className="col-50">Geom Format</label>
                             <div className="col-50">
                                 <select id="geom" name="geom">
                                     <option value="point">Point</option>
@@ -262,18 +254,14 @@ class SchemaConfig extends React.Component {
                                     <option value="multipolygon">Multi Polygon</option>
                                 </select>
                             </div>
-                            <label className="col-50">DateTimeFormat</label>
-                            <div className="col-50">
-                                <select id="dtg" name="dtg" >
-                                    <option value="yyyyMMdd">yyyy/mm/dd</option>
-                                    <option value="ddMMyyyy">dd/mm/yyyy</option>
-                                    <option value="MMddyyyy">mmm/dd/yyyy</option>
-                                </select>
-                            </div>
-                            <label className="col-50">FeatureID</label>
+                            <label className="col-50">Data Source</label>
+                            <input className="col-25" type="text" id="dataSource"></input>
+                            <div className="h7">{this.state.errorMsg["geomConfigurations"]}</div>
+                            <label className="col-50">Feature ID</label>
                             <input className="col-25" type="text" id="featureId"></input>
-                            <div className="h7">{this.state.errorMsg["configurations"]}</div>
+                            <div className="h7">{this.state.errorMsg["geomConfigurations"]}</div>
                         </div>
+
                     </h6>
                     <button className="w3-btn w3-white w3-border w3-border-red w3-round-large" onClick={this.removeFeature}>Remove</button>
                     <button className="w3-btn w3-white w3-border w3-border-green w3-round-large" onClick={this.addFeature}>Add Feature</button>
@@ -286,4 +274,4 @@ class SchemaConfig extends React.Component {
     }
 }
 
-export default SchemaConfig;
+export default BulkIngest;
