+++ src/components/configuration/single/granularity.js
