+++ src/components/query.js
@@ -2,35 +2,35 @@ import React from 'react';
 import axios from 'axios';
 class Queries extends React.Component {
     state = {
-        query:[{queryName:"",catalog:"",schema:"",query:""}],
-        response:""
+        query: [{ queryName: "", catalog: "", schema: "", query: "" }],
+        response: ""
     }
-    
+
     handleChange = (e) => {
         let query = this.state.query
         query[e.target.dataset.id][e.target.id] = e.target.value.toUpperCase()
-        this.setState({query},()=>{
+        this.setState({ query }, () => {
             console.log(query)
         })
     }
 
-    addQuery = () =>{
+    addQuery = () => {
         this.setState((prevState) => ({
-            query: [...prevState.query, {queryName:"",catalog:"",schema:"",query:""}],
+            query: [...prevState.query, { queryName: "", catalog: "", schema: "", query: "" }],
         }));
     }
 
-    postQuery = (e) =>{
+    postQuery = (e) => {
         let response = this.state.response
         axios
-        .post('https://localhost:8080/query', this.state.query)
-        .then(response =>{
-            console.log(response)
-        })
-        .catch(error =>{
-            console.log(error)
-        })
-    } 
+            .post('https://localhost:8080/query', this.state.query)
+            .then(response => {
+                console.log(response)
+            })
+            .catch(error => {
+                console.log(error)
+            })
+    }
 
     removeQuery = (e) => {
         var arrayQuery = this.state.query;
@@ -46,36 +46,36 @@ class Queries extends React.Component {
     }
     render() {
         let query = this.state.query
-        return(
+        return (
             <div >
                 {
-                    query.map((val, idx)=>{
-                        return(
+                    query.map((val, idx) => {
+                        return (
                             <div key={idx} >
-                            <form className="w3-container" onSubmit={this.handleSubmit} onChange={this.handleChange}>
-                            <h6>
-                            <label  className="col-100">Query #{idx+1}</label>
-                                <div >
-                                    <label  className="col-25">Query Name</label>
-                                    <input id="queryName" data-id={idx} className="col-75"/>
-                                    <label  className="col-25">Catalaog</label>
-                                    <input id="catalog" data-id={idx} className="col-75"/>
-                                    <label  className="col-25">Schema</label>
-                                    <input id="schema" data-id={idx} className="col-75 "/>
-                                    <label  className="col-25">Query</label>
-                                    <input id="query" data-id={idx} className="col-75 "/>
-                                </div>
-                            </h6>
-                        </form>
-                        <br />
-                        </div>
+                                <form className="w3-container" onSubmit={this.handleSubmit} onChange={this.handleChange}>
+                                    <h6>
+                                        <label className="col-100">Query #{idx + 1}</label>
+                                        <div >
+                                            <label className="col-25">Query Name</label>
+                                            <input id="queryName" data-id={idx} className="col-75" />
+                                            <label className="col-25">Catalaog</label>
+                                            <input id="catalog" data-id={idx} className="col-75" />
+                                            <label className="col-25">Schema</label>
+                                            <input id="schema" data-id={idx} className="col-75 " />
+                                            <label className="col-25">Query</label>
+                                            <input id="query" data-id={idx} className="col-75 " />
+                                        </div>
+                                    </h6>
+                                </form>
+                                <br />
+                            </div>
                         )
                     })
                 }
                 <button className="w3-button w3-circle w3-teal" onClick={this.removeQuery}>-</button>
                 <button className="w3-button w3-circle w3-teal" onClick={this.addQuery}>+</button>
-                <div className="response w3-panel w3-border">{this.state.response}</div> 
-                <button className="w3-btn w3-white w3-border w3-border-green w3-round-large" onClick={this.postQuery}>Execute</button>  
+                <div className="response w3-panel w3-border">{this.state.response}</div>
+                <button className="w3-btn w3-white w3-border w3-border-green w3-round-large" onClick={this.postQuery}>Execute</button>
             </div>
         )
     }
